Add support to window size change
We have discovered that we don't need the ptys descriptor in the
parent process.  As we have implemented it to get the ptys and
do the window change in the slave pty, it is not neccessary, we
can do it in the master pty and it is replicated in the child.

Signed-off-by: Luis Colorado BaseN <lcu@basen.net>