Add control of input buffer size
The input buffer is stopped (sending back an XOFF)
when it reaches two windows size (or the buffer capacity, if
less) and restarted (sending back an XON) when it reaches less
than one window size.  The reason for this is to control that the
sending process doesn't continue sending data at such high pace
when we cannot output it.  The buffer size is big enough to allow
good working of the XON, XOFF and small to allow quick response
in case an interrupt key is pressed.

Signed-off-by: Luis Colorado <luiscoloradourcola@gmail.com>